requirements.txt
numpy
pandas
tensorflow_cpu
matplotlib
tqdm
fancyimpute
torch
click